Refactor: move senat amendements parsing functions to repondeur
